requirements.txt
google-cloud-aiplatform
flask
flask-cors
python-dotenv
Pillow